# Speed up API generation