--- a/.github/workflows/update_apis.yml
+++ b/.github/workflows/update_apis.yml
@@ -19,7 +19,7 @@ jobs:
       - uses: julia-actions/julia-buildpkg@latest
       - name: UpdateAPI
         # Run against the latest version, since everything merged into master will be tagged.
-        run: julia --project -e 'using AWS; AWS.AWSMetadata.parse_aws_metadata();'
+        run: julia --threads=auto --project -e 'using AWS; AWS.AWSMetadata.parse_aws_metadata();'
         env:
           GITHUB_AUTH: ${{ secrets.GITHUB_TOKEN }}
       - name: FormatCode