--- a/src/api_generation/high_level.jl
+++ b/src/api_generation/high_level.jl
@@ -4,6 +4,16 @@ Generate the `src/services/{service}.jl` file.
 function _generate_high_level_wrapper(
     services::AbstractArray{<:AbstractDict}, repo_name::String, auth::GitHub.OAuth2
 )
+    service_dir = joinpath(@__DIR__, "..", "services")
+
+    # Remove old service files to ensure services that no longer exist are removed.
+    for file in readdir(service_dir)
+        path = joinpath(service_dir, file)
+        if endswith(path, ".jl")
+            rm(path)
+        end
+    end
+
     Threads.@threads for service in services
         service_name = service["path"]
         @info "Generating high level wrapper for $service_name"
@@ -22,7 +32,7 @@ function _generate_high_level_wrapper(
             _generate_high_level_definitions(service_name, protocol, operations, shapes)
         )
 
-        service_path = joinpath(@__DIR__, "..", "services", "$service_name.jl")
+        service_path = joinpath(service_dir, "$service_name.jl")
         open(service_path, "w") do f
             println(
                 f,