--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -3,12 +3,13 @@ using AWS: AWSCredentials
 using AWS: AWSServices
 using AWS.AWSExceptions: AWSException, InvalidFileName, NoCredentials, ProtocolNotDefined
 using AWS.AWSMetadata:
+    ServiceFile,
     _clean_documentation,
     _filter_latest_service_version,
     _generate_low_level_definition,
     _generate_high_level_definition,
     _generate_high_level_definitions,
-    _get_aws_sdk_js_files,
+    _get_service_files,
     _get_service_and_version,
     _get_function_parameters,
     _clean_uri,