In working on some other PRs it was necessary for me to regenerate the service definitions locally. I noticed the process was kind of slow so I looked into some minor improvements and this PR was the outcome.

I started off with adding threading support to the high-level and low-level loops. Here are the resulting timings:

No threading: 1m 45s
Threading (1): 1m 38s
Threading (2): 60s
Threading (4): 37s

I then noticed we were downloading the service definition files twice. To fix this I introduced the `ServiceFile` type which handles fetching of the service definition JSON files and caches the results. This means we only download the service definition files for the low-level API generation. Additionally, this change reduces how many GitHub API call and avoids rate limiting over multiple calls.

Finally, I added in some logic to remove the high-level service definitions before building the new ones. This was done to handle the scenario which AWS actually removes a service definition file.